semantic-kernel-production-blueprint/requirements.txt
semantic-kernel>=1.36.0
fastapi
uvicorn
python-dotenv
pydantic-settings
rich
aiohttp
openai>=1.0.0
